refactor: ErrorException 수정
